requirements_gitlab.txt
astroid==2.6.2
boto3==1.17.103
botocore==1.20.103
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.1
colorama==0.4.4
flake8==3.9.2
idna==2.10
importlib-metadata==4.6.0
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
pylint==2.9.3
python-dateutil==2.8.1
requests==2.26.0
s3transfer==0.4.2
six==1.16.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.6
wrapt==1.11.2
zipp==3.4.1
